Поправлена логика сортировки списка
